Avalanche: reorganise statement and statement simplifiers
